removed ConfigManager from Mony.class and moved it to core module
