+++ API/src/main/java/io/github/invvk/mony/api/Mony.java
@@ -1,6 +1,5 @@
 package io.github.invvk.mony.api;
 
-import io.github.invvk.mony.api.config.ConfigManager;
 import io.github.invvk.mony.api.database.IUserManager;
 import io.github.invvk.mony.api.database.storage.IStorage;
 
@@ -11,8 +10,6 @@ public interface Mony {
     Optional<IStorage> getStorage();
 
     Optional<IUserManager> getUserManager();
-
-    ConfigManager getConfigManager();
 
     boolean isTestEnvironment();
